# docs/runtime-api.full.openapi.yaml
...
openapi: 3.0.3
info:
  title: MisterMorph Runtime API — Full
  version: '2026-10-02'
  description: Full Runtime API reference for shared control/data routes and runtime-mounted management,
    provider login and setup extensions. Availability varies by runtime; Console-specific operations are
    marked. Browser-only /api routes are outside this runtime API. The document revision is not a negotiated
    protocol version. Use runtime-api.control.openapi.yaml for task execution, results and approvals.
    See runtime-api.md for behavior and examples.
servers:
- url: http://127.0.0.1:9080/runtime
  description: Default Console listener after setting server.auth_token. Replace with the full runtime
    base URL, including any console.base_path or reverse-proxy prefix.
security:
- RuntimeBearer: []
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags:
      - Tasks
      summary: List tasks newest first
      parameters:
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/TaskStatus'
        description: Optional status filter; the server also accepts case-insensitive values and an empty
          value for all statuses.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 20
      - name: topic_id
        in: query
        required: false
        schema:
          type: string
      - name: cursor
        in: query
        required: false
        schema:
          type: string
        description: Opaque next_cursor from a previous page. Omit for the newest page; do not construct
          or decode cursors.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      operationId: submitTask
      tags:
      - Tasks
      summary: Submit task text or a runtime command
      description: Returns HTTP 200 after submission, not the final agent result. Some commands and steering
        acknowledgements finish synchronously. Console may steer an active task in the same topic when
        no file references are supplied. Follow steer_target_task_id when present. A runtime without Submit
        returns 405. No idempotency key or automatic retry guarantee. The JSON decoder reads up to 1 MiB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitTaskRequest'
            example:
              task: Summarize the project structure.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitTaskResponse'
              example:
                id: task-example
                status: queued
                topic_id: topic-example
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
  /tasks/{task_id}:
    get:
      operationId: getTask
      tags:
      - Tasks
      summary: Read task status and result
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskInfo'
              example:
                id: task-example
                status: done
                task: Summarize the project structure.
                model: configured-model
                timeout: 2m
                created_at: '2026-10-02T00:00:00Z'
                finished_at: '2026-10-02T00:00:10Z'
                topic_id: topic-example
                result:
                  final:
                    output: The CLI starts the runtime and dispatches tasks to the agent engine.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
      description: Authoritative task state and result endpoint. Poll until done, failed or canceled.
        pending requires an approval decision. HTTP 200 reports successful lookup, not successful execution;
        inspect status and error. Console normally places the answer at result.final.output. WebSocket
        previews do not replace this result.
  /tasks/{task_id}/stop:
    post:
      operationId: stopTask
      tags:
      - Tasks
      summary: Request cancellation by task ID
      description: No request body. Console returns 200 with found=false when no active task matches.
        Cancellation is asynchronous; keep polling.
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopTaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
  /topics:
    get:
      operationId: listTopics
      tags:
      - Topics
      summary: List topics by most recent update
      description: Requires a topic reader. There is no POST /topics; Console creates a topic when a task
        is first submitted without topic_id.
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 100
      - name: cursor
        in: query
        required: false
        schema:
          type: string
        description: Opaque next_cursor from a previous page. Omit for the newest page; do not construct
          or decode cursors.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
  /topics/{topic_id}:
    get:
      operationId: getTopic
      tags:
      - Topics
      summary: Read a topic
      parameters:
      - name: topic_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
    delete:
      operationId: deleteTopic
      tags:
      - Topics
      summary: Delete a topic
      description: Requires a topic deleter. Console also stops topic work and removes associated context.
        A missing or already deleted topic returns 404.
      parameters:
      - name: topic_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '204':
          description: Delete a topic
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /topics/{topic_id}/regenerate-title:
    post:
      operationId: regenerateTopicTitle
      tags:
      - Topics
      summary: Generate a topic title from its history
      description: No request body. Requires a title generator. Reserved default and _awareness topics
        return 400. A busy topic or concurrent title change returns 409.
      parameters:
      - name: topic_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
        '409':
          $ref: '#/components/responses/Conflict'
  /topics/{topic_id}/tags:
    put:
      operationId: setTopicTags
      tags:
      - Topics
      summary: Replace a topic's tags
      description: Console only. Replaces the whole tag list; send an empty list to clear it. Each tag is
        trimmed and its inner whitespace folded; empty tags and case-insensitive duplicates are dropped,
        keeping the first spelling. At most 5 tags of up to 32 characters, plus the reserved tag pinned. Does not change updated_at.
        JSON body limit is 64 KiB.
      parameters:
      - name: topic_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
              required:
              - tags
      responses:
        '200':
          description: The updated topic.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /todo/tasks:
    get:
      operationId: listScheduledTasks
      tags:
      - Schedules
      summary: Read scheduled and system tasks
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoTasks'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replaceScheduledTasks
      tags:
      - Schedules
      summary: Replace the full scheduled task list
      description: Not a patch. Missing or null tasks writes an empty list. The decoder reads up to 4
        MiB. System tasks are returned separately and are not replaced here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tasks:
                  type: array
                  items:
                    $ref: '#/components/schemas/CronTask'
                  nullable: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoTasks'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
  /todo/tasks/{id}/run:
    post:
      operationId: runScheduledTask
      tags:
      - Schedules
      summary: Trigger a saved scheduled task now
      description: No request body. Requires a cron runner. A busy runner returns 409; a missing saved
        schedule returns 404.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  id:
                    type: string
                  triggered_at:
                    type: string
                    format: date-time
                required:
                - ok
                - id
                - triggered_at
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /settings/agent/test:
    post:
      operationId: testLLMConnection
      tags:
      - Management
      summary: Test LLM connection settings
      description: Available when agent settings are enabled; Console uses its own settings owner. Makes
        inference requests using the selected profile and supplied LLM settings. Does not save them. JSON
        body limit is 1 MiB.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  provider:
                    type: string
                  api_base:
                    type: string
                  model:
                    type: string
                  benchmarks:
                    type: array
                    items:
                      $ref: '#/components/schemas/LLMBenchmark'
                    nullable: true
        '400':
          description: Invalid request or configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '401':
          description: Missing or invalid runtime token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '405':
          description: Unsupported method or read-only settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '500':
          description: Storage or internal failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '502':
          description: Remote provider or endpoint request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '503':
          description: Runtime, settings writer or secret store unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionTestRequest'
...
components:
  securitySchemes:
    RuntimeBearer:
      type: http
      scheme: bearer
      description: The configured server.auth_token. This is not a Console browser session token.
  responses:
    BadRequest:
      description: Invalid JSON, parameter or runtime request.
      content:
        text/plain:
          schema:
            type: string
    Unauthorized:
      description: Missing or invalid Authorization Bearer token.
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource or route not found.
      content:
        text/plain:
          schema:
            type: string
    MethodNotAllowed:
      description: Unsupported method; task submission can also be disabled.
      content:
        text/plain:
          schema:
            type: string
    Conflict:
      description: Concurrent change or runtime busy.
      content:
        text/plain:
          schema:
            type: string
...
    Unavailable:
      description: Required runtime capability is absent, queue is full, or execution/storage is unavailable.
      content:
        text/plain:
          schema:
            type: string
  schemas:
    TaskStatus:
      type: string
      enum:
      - queued
      - running
      - pending
      - done
      - failed
      - canceled
      description: pending waits for approval. done, failed and canceled are terminal task states.
...
    TaskInfo:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        task:
          type: string
        model:
          type: string
        llm_profile:
          type: string
        timeout:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        pending_at:
          type: string
          format: date-time
        resumed_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        approval_request_id:
          type: string
        error:
          type: string
        result:
          description: Runtime-dependent JSON value. Console commonly stores the answer in result.final.output,
            alongside retained execution details. output may itself be structured JSON. Do not assume
            a plain string or require a fixed result envelope.
        topic_id:
          type: string
        steer_target_task_id:
          type: string
        file_references:
          type: array
          items:
            $ref: '#/components/schemas/FileReference'
          nullable: true
        conversation:
          $ref: '#/components/schemas/TaskConversation'
      required:
      - id
      - status
      - task
      - model
      - timeout
      - created_at
      x-source: internal/taskdomain/domain.go:TaskInfo
...
    TopicInfo:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        icon:
          type: string
        title_customized:
          type: boolean
        tags:
          type: array
          maxItems: 6
          items:
            type: string
            minLength: 1
            maxLength: 32
          description: Labels set with PUT /topics/{topic_id}/tags. The reserved tag pinned (stored in
            lower case) pins the topic to the top of its list; it does not count toward the five-tag
            limit. Absent when the topic has none.
        title_revision:
          type: integer
        llm_title_generated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
      required:
      - id
      - created_at
      - updated_at
      x-source: internal/taskdomain/domain.go:TopicInfo
    SubmitTaskRequest:
      type: object
      properties:
        task:
          type: string
          minLength: 1
          description: Nonblank task text, or a supported runtime slash command.
        model:
          type: string
          description: Runtime-specific model override. For persistent Console profile selection use the
            /models command in task text.
        llm_profile:
          type: string
          description: Configured LLM profile for this submission; accepted behavior depends on the runtime.
        timeout:
          type: string
          description: Positive Go duration, for example 30s or 2m. Omit to use the runtime default.
        topic_id:
          type: string
          description: 'Console: omit to create a topic on first submission; otherwise use an existing
            topic ID. An unknown explicit ID is rejected. The reserved default topic is also accepted.'
        topic_title:
          type: string
          description: Optional initial topic title.
        workspace_dir:
          type: string
          description: Existing directory on the server. Console may attach it to the selected or new
            topic.
        file_references:
          type: array
          items:
            $ref: '#/components/schemas/FileReference'
          nullable: true
        trigger:
          $ref: '#/components/schemas/TaskTrigger'
      required:
      - task
      x-source: internal/daemonruntime/types.go:SubmitTaskRequest
    SubmitTaskResponse:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        topic_id:
          type: string
        steer_target_task_id:
          type: string
          description: When present, this submission records steering input for this active task. Continue
            tracking that task for the main result.
      required:
      - id
      - status
      x-source: internal/daemonruntime/types.go:SubmitTaskResponse
    StopTaskResponse:
      type: object
      properties:
        status:
          type: string
          description: Console returns stopping or not_found. stopping acknowledges cancellation; poll
            the task for its final state.
        found:
          type: boolean
        task_id:
          type: string
        topic_id:
          type: string
        progress:
          type: string
        message:
          type: string
      required:
      - status
      - found
      x-source: internal/daemonruntime/types.go:StopTaskResponse
...
    TaskPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskInfo'
          nullable: true
        limit:
          type: integer
        next_cursor:
          type: string
        has_next:
          type: boolean
      required:
      - items
      - limit
      - has_next
    TopicPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TopicInfo'
          nullable: true
        limit:
          type: integer
        next_cursor:
          type: string
        has_next:
          type: boolean
      required:
      - items
      - limit
      - has_next
...
    TodoTasks:
      type: object
      properties:
        ok:
          type: boolean
        version:
          type: integer
        exists:
          type: boolean
        task_count:
          type: integer
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/CronTask'
          nullable: true
        system_tasks:
          type: array
          items:
            $ref: '#/components/schemas/CronTask'
          nullable: true
        heartbeat_enabled:
          type: boolean
        llm_default_route:
          $ref: '#/components/schemas/LLMProfile'
        llm_profiles:
          type: array
          items:
            $ref: '#/components/schemas/LLMProfile'
          nullable: true
        chat_options:
          type: array
          items:
            $ref: '#/components/schemas/TodoChatOption'
          nullable: true
...